+++ src/pyramid_bimt/const.py
@@ -4,6 +4,12 @@
 from flufl.enum import Enum
 
 
+class Modes(Enum):
+    development = 'development'
+    testing = 'testing'
+    production = 'production'
+
+
 class BimtPermissions(Enum):
     view = 'View'
     manage = 'Manage'